+++ workspace/examples/CHANGELOG.md
@@ -5,6 +5,13 @@
 ### Patch Changes
 
 - Updated dependencies
+  - @creative-introvert/prediction-testing@0.0.4
+
+## null
+
+### Patch Changes
+
+- Updated dependencies
   - @creative-introvert/prediction-testing@0.0.3
 
 ## null